requirements.txt
cycler==0.10.0
fvcore==0.1.5.post20210722
iopath==0.1.9
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.21.1
pandas==1.3.0
Pillow==8.3.1
portalocker==2.3.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
sklearn==0.0
tabulate==0.8.9
termcolor==1.1.0
threadpoolctl==2.2.0
torch==1.7.1
torchaudio==0.7.2
torchvision==0.8.2
tqdm==4.61.2
typing-extensions==3.10.0.0
yacs==0.1.8
